+++ example/group.py
@@ -2,8 +2,8 @@ import yh_user_sdk as sdk
 # from yh_user_sdk.core.group import group
 # 如果你想要单独导入group的话请使用上面这个,会导致下列命令请求格式变化 传入token的话直接group("token")就行
 
-# 设置token
-client = sdk.set_token("你的token")
+# 设置token, timeout参数可不写
+client = sdk.set_token("你的token", timeout = 114514)
 
 # 获取群聊信息
 print(client.group.info("big"))