+++ changelog.md
@@ -1,7 +1,13 @@
+### 0.1.10
+修正 bot 的 proto 中 avatar_id 类型.  
+新增 sticker, 目前支持 list, add, remove, sort, detail.  
+新增 expression, 目前支持 list, create, add, delete, topping.  
+新增 sticky, 目前支持 list, add, delete, top  
+
 ### 0.1.9
 修正 msg 的 button_report 对按钮值的变量命名.  
-新增 bot 的 info, bot_group_list  
-新增 group.tag 的 member(s)
+新增 bot 的 info, bot_group_list.  
+新增 group.tag 的 member(s).
 
 ### 0.1.8
 更新群云盘相关接口的SDK.